+++ misc/get_estimates_art_coverage_participants_vl200.R
@@ -313,8 +313,15 @@ stats[['max_rhat']] = convergence[, round(max(rhat), 4)]
 
 #########
 
-file.name <- file.path(gitdir.fit,'RCCS_art_posterior_samples_vl200_221208.rds')
-saveRDS(nsinf.samples, file = file.name)
+# file.name <- file.path(gitdir.fit,'RCCS_art_posterior_samples_vl200_221208.rds')
+file.name <- file.selfreportedart.vl200 
+if(! file.exists(file.name))
+{
+    cat("Saving file:", file.name, '\n')
+    saveRDS(nsinf.samples, file = file.name)
+}else{
+    cat("File:", file.name, "already exists...\n")
+}
 
 file.name <- file.path(outdir, 'RCCS_art_model_fit_vl200_221208.RDS')
 saveRDS(stats, file = file.name)
