+++ misc/get_estimates_unsuppressed_proportion_participants_vl200.R
@@ -292,8 +292,15 @@ stats[['max_rhat']] = convergence[, round(max(rhat), 4)]
 
 #########
 
-file.name <- file.path(gitdir, 'fit', paste0('RCCS_nonsuppressed_proportion_posterior_samples_vl_200_221121.rds'))
-saveRDS(nsinf.samples, file = file.name)
+# file.name <- file.path(gitdir.fit,'RCCS_nonsuppressed_proportion_posterior_samples_vl_200_221121.rds')
+file.name <- file.unsuppressedviralload.vl200
+if(! file.exists(file.name))
+{
+    cat("Saving file:", file.name, '\n')
+    saveRDS(nsinf.samples, file = file.name)
+}else{
+    cat("File:", file.name, "already exists...\n")
+}
 
-file.name <- file.path(outdir, paste0('RCCS_nonsuppressed_proportion_model_fit_vl200_221121.RDS'))
+file.name <- file.path(outdir,'RCCS_nonsuppressed_proportion_model_fit_vl200_221121.RDS')
 saveRDS(stats, file = file.name)
