+++ scripts_for_confidential_data/get_sample_collection_dates_and_randomize.R
@@ -5,22 +5,12 @@ require(here)
 # PATHS # 
 #########
 
-usr <- Sys.info()[['user']]
-indir <- here::here()
-
-if(usr == 'andrea')
-{
-    indir.deepsequencedata <- '/home/andrea/HPC/project/ratmann_pangea_deepsequencedata/live'
-    indir.deepanalyses_xiaoyue <- '/home/andrea/HPC/project/ratmann_xiaoyue_jrssc2022_analyses/live/PANGEA2_RCCS1519_UVRI'
-}
-
-# TODO: better on HPC, as anyways do not push this
-outdir.confidential <- file.path(indir.deepsequencedata, 'RCCS_R15_R18')
-outdir.data <- file.path(indir, 'data')
+gitdir <- here()
+source(file.path(gitdir, 'paths.R'))
 
 # path from phylo analyses
-path.anonymisation.keys <- file.path(indir.deepanalyses_xiaoyue, 'important_anonymisation_keys_210119.csv')
-path.selected.samples <- file.path(indir.deepanalyses_xiaoyue,"210120_RCCSUVRI_phscinput_samples.rds" )
+path.anonymisation.keys <- file.path(indir.deepanalyses.xiaoyue, 'important_anonymisation_keys_210119.csv')
+path.selected.samples <- file.path(indir.deepanalyses.xiaoyue,"210120_RCCSUVRI_phscinput_samples.rds" )
 
 # path from PANGEA data
 path.sdates.rccs <- file.path(indir.deepsequencedata, 'PANGEA2_RCCS', '200316_pangea_db_sharing_extract_rakai.csv')
@@ -65,8 +55,14 @@ ddates <- unique(ddates) |>
     subset(select=c('rename_id','aid' ,'visit_dt'))
 
 catn("Save the original in encrypted data folder")
-filename <- file.path(outdir.confidential, 'sequences_collection_dates.rds')
-saveRDS(ddates, file=filename)
+filename <- path.collection.dates.confidential
+if(! file.exists(filename))
+{
+    cat("Saving file:", filename, '\n')
+    saveRDS(ddates, file=filename)
+}else{
+    cat("File:", filename, "already exists...\n")
+}
 
 catn("Perturb dates in (-3, 3) months")
 ddates[, visit_dt_perturbed := perturb_dates(visit_dt, -.25, .25) ]
@@ -76,6 +72,12 @@ ddates_perturbed <- ddates[, .(
     visit_dt = visit_dt_perturbed
 )]
 
-catn("Save the perturbed dates in the encrypted data folder")
-filename <- file.path(outdir.data, 'sequences_collection_dates_randomized.rds')
-saveRDS(ddates_perturbed, file=filename)
+catn("Save the perturbed dates in the zenodo dir")
+filename <- path.collection.dates.randomized 
+if(! file.exists(filename))
+{
+    cat("Saving file:", filename, '\n')
+    saveRDS(ddates_perturbed, file=filename)
+}else{
+    cat("File:", filename, "already exists...\n")
+}
